az/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
docopt==0.6.2
    # via -r requirements.in
duckdb==1.3.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
